Execute simulations defined in a optic_simulation object

Execute simulations defined in a optic_simulation object

Usage

dispatch_simulations(object, seed = NULL, use_future = FALSE, verbose = 0, ...)

Arguments

object

Simulation scenarios object created using optic_simulation

seed

Specified as either NULL or a numeric. Sets a seed, which is becomes an index in results, for each independent set of simulations in optic_simulation.

use_future

Runs simulation scenarios in parallel. Default FALSE, set to TRUE if you have already setup a future plan (e.g., multiprocess, cluster, etc) and would like for the iterations to be run in parallel.

verbose

Default TRUE. IF TRUE, provides details on what's currently running.

...

additional parameters to be passed to future_apply. User can pass future.globals and future.packages if your code relies on additional packages

Value

A list of dataframes, where each list entry contains results for a set of simulation parameters, with dataframes containing estimated treatment effects and summary statistics by model and draw.

Examples

# Set up a basic model and simulation scenario:
data(overdoses)

eff <- 0.1*mean(overdoses$crude.rate, na.rm = T)
form <- formula(crude.rate ~ state + year + population + treatment_level)
mod <- optic_model(name = 'lin', 
                   type = 'reg', 
                   call = 'lm', 
                   formula = form, 
                   se_adjust = 'none')

sim <- optic_simulation(x = overdoses, 
                        models = list(mod), 
                        method = 'no_confounding', 
                        unit_var = 'state', 
                        treat_var = 'state',
                        time_var = 'year', 
                        effect_magnitude = list(eff), 
                        n_units = 10, 
                        effect_direction = 'pos', 
                        iters = 10,
                        policy_speed = 'instant', 
                        n_implementation_periods = 1)
#> Number of Simulations: 1
#> Number of Models: 1
#> Iteration per Simulation : 10
#> Total number of Iterations to Run: 10

# Finally, dispatch the simulation:
dispatch_simulations(sim)
#> [[1]]
#>       outcome se_adjustment   estimate        se  variance     t_stat
#> 1  crude.rate          none  2.2230143 0.6290334 0.3956830  3.5340163
#> 2  crude.rate          none  2.6394772 0.5909810 0.3492586  4.4662640
#> 3  crude.rate          none  0.6129511 0.6009056 0.3610876  1.0200454
#> 4  crude.rate          none -0.5662105 0.6342965 0.4023320 -0.8926591
#> 5  crude.rate          none  1.5222472 0.6116178 0.3740763  2.4888864
#> 6  crude.rate          none -0.8285078 0.6259276 0.3917853 -1.3236481
#> 7  crude.rate          none  4.2845626 0.6035313 0.3642500  7.0991560
#> 8  crude.rate          none  2.0591920 0.6248139 0.3903924  3.2956887
#> 9  crude.rate          none -0.3222015 0.6265473 0.3925616 -0.5142493
#> 10 crude.rate          none  3.6770233 0.5841235 0.3412003  6.2949417
#>         p_value      mse model_name model_call
#> 1  4.299929e-04 11.00617        lin         lm
#> 2  8.964756e-06 10.96886        lin         lm
#> 3  3.079792e-01 11.02002        lin         lm
#> 4  3.722768e-01 10.93375        lin         lm
#> 5  1.299294e-02 11.03219        lin         lm
#> 6  1.859538e-01 10.89975        lin         lm
#> 7  2.534639e-12 10.73769        lin         lm
#> 8  1.019936e-03 11.01471        lin         lm
#> 9  6.072031e-01 10.95673        lin         lm
#> 10 4.785480e-10 10.83052        lin         lm
#>                                               model_formula policy_speed
#> 1  crude.rate ~ state + year + population + treatment_level      instant
#> 2  crude.rate ~ state + year + population + treatment_level      instant
#> 3  crude.rate ~ state + year + population + treatment_level      instant
#> 4  crude.rate ~ state + year + population + treatment_level      instant
#> 5  crude.rate ~ state + year + population + treatment_level      instant
#> 6  crude.rate ~ state + year + population + treatment_level      instant
#> 7  crude.rate ~ state + year + population + treatment_level      instant
#> 8  crude.rate ~ state + year + population + treatment_level      instant
#> 9  crude.rate ~ state + year + population + treatment_level      instant
#> 10 crude.rate ~ state + year + population + treatment_level      instant
#>    n_implementation_periods prior_control effect_magnitude n_units
#> 1                         1         level         1.265073      10
#> 2                         1         level         1.265073      10
#> 3                         1         level         1.265073      10
#> 4                         1         level         1.265073      10
#> 5                         1         level         1.265073      10
#> 6                         1         level         1.265073      10
#> 7                         1         level         1.265073      10
#> 8                         1         level         1.265073      10
#> 9                         1         level         1.265073      10
#> 10                        1         level         1.265073      10
#>    effect_direction  n mean_es_prior max_es_prior mean_es_outcome
#> 1               pos 10   -0.62202040    0.9252669      -0.3696761
#> 2               pos 10    0.49183037    0.7409951       0.5726075
#> 3               pos 10    0.23933609    0.3442886       0.7006378
#> 4               pos 10   -0.48115781    0.5391861      -0.2567397
#> 5               pos 10    0.29544392    0.3848879       0.5602086
#> 6               pos 10    0.46832529    0.8857470       0.6061811
#> 7               pos 10    0.95142132    1.0192945       1.0762423
#> 8               pos 10    0.59287722    1.1974957       0.9066567
#> 9               pos 10   -0.08768851    0.3515408      -0.1214716
#> 10              pos 10    0.32551729    0.4359520       0.4183461
#>    max_es_outcome n_unique_enact_years mu1_prior mu0_prior sd_prior
#> 1       0.8231403                    9  11.22509  12.14383 5.098109
#> 2       0.9682922                    6  14.35789  11.65420 5.098109
#> 3       1.7288473                    8  15.14690  11.66513 5.098109
#> 4       0.7625312                    6  11.90702  12.05505 5.098109
#> 5       1.0601872                    8  15.08235  11.67761 5.098109
#> 6       1.2945110                    7  15.86296  11.66129 5.098109
#> 7       1.3064672                    7  15.84848  11.36242 5.098109
#> 8       1.8188744                    9  15.95766  11.64018 5.098109
#> 9       1.3508565                    7  13.73750  11.82863 5.098109
#> 10      0.8655672                    7  14.57619  11.71356 5.098109
#>    mu1_prior_old mu0_prior_old sd_prior_old      mu1      mu0       sd iter
#> 1       11.22509      12.14383     5.098109 14.51851 12.57866 6.668072    1
#> 2       14.35789      11.65420     5.098109 18.09935 12.08026 6.753155    2
#> 3       15.14690      11.66513     5.098109 18.69366 12.17300 6.737871    3
#> 4       11.90702      12.05505     5.098109 14.45080 12.58009 6.669050    4
#> 5       15.08235      11.67761     5.098109 18.65530 12.18297 6.734333    5
#> 6       15.86296      11.66129     5.098109 18.41833 12.27984 6.715858    6
#> 7       15.84848      11.36242     5.098109 20.07735 11.75483 6.806116    7
#> 8       15.95766      11.64018     5.098109 19.63716 12.16704 6.738448    8
#> 9       13.73750      11.82863     5.098109 16.32560 12.38616 6.700500    9
#> 10      14.57619      11.71356     5.098109 18.34347 12.18113 6.735275   10
#>